fix: Exported types (#622)
