+++ package.json
@@ -34,7 +34,6 @@
     "serverless": "^3.25.1",
     "ts-jest": "^27.1.5",
     "ts-node": "^10.9.1",
-    "ts-toolbelt": "^9.6.0",
     "typescript": "^4.9.3"
   },
   "dependencies": {
